+++ src/flows/Termination/tests/TerminationFlow.test.tsx
@@ -8,8 +8,14 @@ import {
   RenderProps,
   TerminationFlow,
 } from '@/src/flows/Termination/TerminationFlow';
-import { render, screen, waitFor, within } from '@testing-library/react';
+import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
+import {
+  fillCheckbox,
+  fillRadio,
+  fillSelect,
+  selectDayInCalendar,
+} from '@/src/common/testHelpers';
 
 const queryClient = new QueryClient();
 
@@ -73,49 +79,6 @@ describe('TerminationFlow', () => {
   afterEach(() => {
     vi.clearAllMocks();
   });
-
-  // TODO: move this to a helper function
-  async function fillRadio(radioName: string, radioValue: string) {
-    const user = userEvent.setup();
-    const radioGroup = screen.getByRole('radiogroup', {
-      name: new RegExp(radioName, 'i'),
-    });
-    expect(radioGroup).toBeInTheDocument();
-
-    const radioButton = within(radioGroup).getByRole('radio', {
-      name: new RegExp(radioValue, 'i'),
-    });
-    expect(radioButton).toBeInTheDocument();
-
-    await user.click(radioButton);
-  }
-
-  // TODO: move this to a helper function
-  async function fillSelect(selectName: string, selectValue: string) {
-    const user = userEvent.setup();
-    const select = screen.getByRole('combobox', {
-      name: new RegExp(selectName, 'i'),
-    });
-    expect(select).toBeInTheDocument();
-
-    await user.click(select);
-
-    const option = screen.getByRole('option', {
-      name: new RegExp(selectValue, 'i'),
-    });
-    expect(option).toBeInTheDocument();
-
-    await user.click(option);
-  }
-
-  async function fillCheckbox(checkboxName: string) {
-    const user = userEvent.setup();
-    const checkbox = screen.getByRole('checkbox', {
-      name: new RegExp(checkboxName, 'i'),
-    });
-    expect(checkbox).toBeInTheDocument();
-    await user.click(checkbox);
-  }
 
   function getYearMonthDate(date: Date) {
     const year = date.getFullYear();
@@ -202,20 +165,7 @@ describe('TerminationFlow', () => {
       await fillCheckbox(values?.riskAssessmentReason);
     }
     if (values?.proposedTerminationDate) {
-      const datePickerButton = screen.getByTestId('date-picker-button');
-      await user.click(datePickerButton);
-      await waitFor(() => {
-        expect(screen.getByRole('dialog')).toBeInTheDocument();
-      });
-      const calendar = screen.getByRole('dialog');
-      expect(calendar).toBeInTheDocument();
-      const dateButton = screen.getByRole('button', {
-        name: new RegExp(values.proposedTerminationDate, 'i'),
-      });
-      await user.click(dateButton);
-      waitFor(() => {
-        expect(screen.queryByRole('dialog')).not.toBeInTheDocument();
-      });
+      await selectDayInCalendar(values?.proposedTerminationDate);
     }
   }
